+++ src/index.ts
@@ -13,6 +13,7 @@ import {
 } from 'git-smart-proxy-core';
 import * as HttpCodes from 'http-status';
 import { Context, Middleware } from 'koa';
+import { Signal } from 'micro-signals';
 import { resolve } from 'path';
 import { Readable, Writable } from 'stream';
 import { promisify } from 'util';
@@ -33,6 +34,12 @@ export const SymbolSource = Symbol('source');
 export class GitSmartProxy {
   public repository?: string;
   public metadata: GitMetadata;
+
+  // Signals
+  public readonly onAccept: Signal<string>;
+  public readonly onReject: Signal<{reason: string; status: number}>;
+  public readonly onFinal: Signal<void>;
+  public readonly onError: Signal<Error>;
 
   // @ts-ignore suppress error [1166]
   private [SymbolSource]?: GitBasePack;
@@ -57,6 +64,12 @@ export class GitSmartProxy {
     this.__context = context;
     this.__service = service;
     this.__status = RequestStatus.PENDING;
+
+    // Signals
+    this.onAccept = new Signal();
+    this.onReject = new Signal();
+    this.onFinal = new Signal();
+    this.onError = new Signal();
 
     if (this.__service === ServiceType.UNKNOWN) {
       return this;
@@ -110,6 +123,7 @@ export class GitSmartProxy {
 
     if (!repo_path) {
       if (!this.repository) {
+        this.onError.dispatch(new Error(''));
         return;
       }
 
@@ -118,7 +132,12 @@ export class GitSmartProxy {
 
     const source = this[SymbolSource];
 
-    await source.accept(repo_path);
+    try {
+      await source.accept(repo_path);
+    } catch (err) {
+      this.onError.dispatch(err);
+      return;
+    }
 
     const ctx = this.__context;
 
@@ -126,6 +145,9 @@ export class GitSmartProxy {
     ctx.type = this.__content_type;
     ctx.status = HttpCodes.OK;
     ctx.body = source;
+
+    this.onAccept.dispatch(repo_path);
+    this.onFinal.dispatch(undefined);
   }
 
   public async reject(): Promise<void>;
@@ -167,6 +189,9 @@ export class GitSmartProxy {
     ctx.type = 'text/plain';
     ctx.status = status;
     ctx.body = reason;
+
+    this.onReject.dispatch({reason, status});
+    this.onFinal.dispatch(undefined);
   }
 
   public async exists(): Promise<boolean>;
@@ -180,7 +205,13 @@ export class GitSmartProxy {
       repo_path = this.repository;
     }
 
-    return exists(this.__command, repo_path);
+    try {
+      return exists(this.__command, repo_path);
+    } catch (err) {
+      this.onError.dispatch(err);
+
+      return false;
+    }
   }
 
   public verbose(...messages: Array<string | Buffer>) {
@@ -239,9 +270,11 @@ export class GitSmartProxy {
       };
     }
 
-    return async(context, next) => {
+    return async(context: Context, next) => {
       // Add proxy to context
       const proxy = context.state[key_name] = await GitSmartProxy.create(context, command);
+
+      proxy.onError.add((err) => context.throw(err));
 
       await next();
 
@@ -264,13 +297,9 @@ export class GitSmartProxy {
   }
 }
 
-export function middleware(options?: MiddlewareOptions) {
-  return GitSmartProxy.middleware(options);
-}
+export const middleware = GitSmartProxy.middleware;
 
-export function create(context: Context, command: GitCommand) {
-  return GitSmartProxy.create(context, command);
-}
+export const create = GitSmartProxy.create;
 
 export interface MiddlewareOptions {
   /**